Migrate HintException to OCP
