+++ apps/federatedfilesharing/lib/AddressHandler.php
@@ -25,8 +25,8 @@
  */
 namespace OCA\FederatedFileSharing;
 
-use OC\HintException;
 use OCP\Federation\ICloudIdManager;
+use OCP\HintException;
 use OCP\IL10N;
 use OCP\IURLGenerator;
