+++ apps/federation/lib/Controller/SettingsController.php
@@ -23,10 +23,10 @@
  */
 namespace OCA\Federation\Controller;
 
-use OC\HintException;
 use OCA\Federation\TrustedServers;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\DataResponse;
+use OCP\HintException;
 use OCP\IL10N;
 use OCP\IRequest;
