+++ apps/federation/lib/Middleware/AddServerMiddleware.php
@@ -28,12 +28,12 @@
  */
 namespace OCA\Federation\Middleware;
 
-use OC\HintException;
 use OCA\Federation\Controller\SettingsController;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\JSONResponse;
 use OCP\AppFramework\Middleware;
+use OCP\HintException;
 use OCP\IL10N;
 use OCP\ILogger;
