+++ apps/federation/tests/Middleware/AddServerMiddlewareTest.php
@@ -26,10 +26,10 @@
  */
 namespace OCA\Federation\Tests\Middleware;
 
-use OC\HintException;
 use OCA\Federation\Controller\SettingsController;
 use OCA\Federation\Middleware\AddServerMiddleware;
 use OCP\AppFramework\Http;
+use OCP\HintException;
 use OCP\IL10N;
 use OCP\ILogger;
 use Test\TestCase;
