+++ lib/private/Config.php
@@ -38,6 +38,8 @@
  */
 namespace OC;
 
+use OCP\HintException;
+
 /**
  * This class is responsible for reading and writing config.php, the very basic
  * configuration file of Nextcloud.
