+++ lib/private/DatabaseSetupException.php
@@ -22,5 +22,7 @@
  */
 namespace OC;
 
+use OCP\HintException;
+
 class DatabaseSetupException extends HintException {
 }
