+++ lib/private/HintException.php
@@ -1,28 +1,28 @@
 <?php
+
+declare(strict_types=1);
 /**
- * @copyright Copyright (c) 2016, ownCloud, Inc.
+ * @copyright Copyright (c) 2021 Gary Kim <gary@garykim.dev>
  *
- * @author Arthur Schiwon <blizzz@arthur-schiwon.de>
- * @author Bart Visscher <bartv@thisnet.nl>
- * @author Lukas Reschke <lukas@statuscode.ch>
- * @author Michael Gapczynski <GapczynskiM@gmail.com>
- * @author Thomas Müller <thomas.mueller@tmit.eu>
+ * @author Gary Kim <gary@garykim.dev>
  *
- * @license AGPL-3.0
+ * @license GNU AGPL version 3 or any later version
  *
- * This code is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License, version 3,
- * as published by the Free Software Foundation.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
  *
- * You should have received a copy of the GNU Affero General Public License, version 3,
- * along with this program. If not, see <http://www.gnu.org/licenses/>
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
+
 namespace OC;
 
 /**
@@ -31,48 +31,7 @@ namespace OC;
  * An Exception class with the intention to be presented to the end user
  *
  * @package OC
+ * @depreacted 23.0.0 Use \OCP\HintException
  */
-class HintException extends \Exception {
-	private $hint;
-
-	/**
-	 * HintException constructor.
-	 *
-	 * @param string $message  The error message. It will be not revealed to the
-	 *                         the user (unless the hint is empty) and thus
-	 *                         should be not translated.
-	 * @param string $hint     A useful message that is presented to the end
-	 *                         user. It should be translated, but must not
-	 *                         contain sensitive data.
-	 * @param int $code
-	 * @param \Exception|null $previous
-	 */
-	public function __construct($message, $hint = '', $code = 0, \Exception $previous = null) {
-		$this->hint = $hint;
-		parent::__construct($message, $code, $previous);
-	}
-
-	/**
-	 * Returns a string representation of this Exception that includes the error
-	 * code, the message and the hint.
-	 *
-	 * @return string
-	 */
-	public function __toString() {
-		return __CLASS__ . ": [{$this->code}]: {$this->message} ({$this->hint})\n";
-	}
-
-	/**
-	 * Returns the hint with the intention to be presented to the end user. If
-	 * an empty hint was specified upon instatiation, the message is returned
-	 * instead.
-	 *
-	 * @return string
-	 */
-	public function getHint() {
-		if (empty($this->hint)) {
-			return $this->message;
-		}
-		return $this->hint;
-	}
+class HintException extends \OCP\HintException {
 }
