+++ lib/private/legacy/OC_Hook.php
@@ -81,7 +81,7 @@ class OC_Hook {
 	 * @param string $signalName name of signal
 	 * @param mixed $params default: array() array with additional data
 	 * @return bool true if slots exists or false if not
-	 * @throws \OC\HintException
+	 * @throws \OCP\HintException
 	 * @throws \OC\ServerNotAvailableException Emits a signal. To get data from the slot use references!
 	 *
 	 * TODO: write example
@@ -107,7 +107,7 @@ class OC_Hook {
 			} catch (Exception $e) {
 				self::$thrownExceptions[] = $e;
 				\OC::$server->getLogger()->logException($e);
-				if ($e instanceof \OC\HintException) {
+				if ($e instanceof \OCP\HintException) {
 					throw $e;
 				}
 				if ($e instanceof \OC\ServerNotAvailableException) {
