+++ lib/public/Encryption/Exceptions/GenericEncryptionException.php
@@ -26,7 +26,7 @@
  */
 namespace OCP\Encryption\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class GenericEncryptionException
