+++ lib/public/Federation/Exceptions/ActionNotSupportedException.php
@@ -22,7 +22,7 @@
  */
 namespace OCP\Federation\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class ActionNotSupportedException
