+++ lib/public/Federation/Exceptions/AuthenticationFailedException.php
@@ -22,7 +22,7 @@
  */
 namespace OCP\Federation\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class AuthenticationFailedException
