+++ lib/public/Federation/Exceptions/BadRequestException.php
@@ -23,7 +23,7 @@
  */
 namespace OCP\Federation\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class BadRequestException
