+++ lib/public/Federation/Exceptions/ProviderAlreadyExistsException.php
@@ -24,7 +24,7 @@
  */
 namespace OCP\Federation\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class ProviderAlreadyExistsException
