+++ lib/public/Federation/Exceptions/ProviderDoesNotExistsException.php
@@ -22,7 +22,7 @@
  */
 namespace OCP\Federation\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class ProviderDoesNotExistsException
