+++ lib/public/Share/Exceptions/GenericShareException.php
@@ -24,7 +24,7 @@
  */
 namespace OCP\Share\Exceptions;
 
-use OC\HintException;
+use OCP\HintException;
 
 /**
  * Class GenericEncryptionException
