+++ tests/Core/Controller/ChangePasswordControllerTest.php
@@ -22,11 +22,11 @@
 
 namespace Tests\Core\Controller;
 
-use OC\HintException;
 use OC\User\Session;
 use OCA\Settings\Controller\ChangePasswordController;
 use OCP\App\IAppManager;
 use OCP\AppFramework\Http\JSONResponse;
+use OCP\HintException;
 use OCP\IGroupManager;
 use OCP\IL10N;
 use OCP\IRequest;
