+++ tests/lib/User/DatabaseTest.php
@@ -22,10 +22,10 @@
 
 namespace Test\User;
 
-use OC\HintException;
 use OC\User\User;
 use OCP\EventDispatcher\Event;
 use OCP\EventDispatcher\IEventDispatcher;
+use OCP\HintException;
 use OCP\Security\Events\ValidatePasswordPolicyEvent;
 use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -82,9 +82,9 @@ class DatabaseTest extends Backend {
 		$this->assertSame($user, $this->backend->checkPassword($user, 'newpass'));
 	}
 
-	
+
 	public function testVerifyPasswordEventFail() {
-		$this->expectException(\OC\HintException::class);
+		$this->expectException(\OCP\HintException::class);
 		$this->expectExceptionMessage('password change failed');
 
 		$user = $this->getUser();
